Ports without indexes now get messages from all indexed output
A few changes to support ports without indexes in a defined manner.
If a component developer sends to an output port without an index
(i.e. output_port.send_message instead of
output_port['key'].send_message, this will result in the same message
being transmitted to all indexed connections for that output port,
including those with nil/no index.  Similarly, if an application
developer connects an input port of a component to the unindexed
output port of another (i.e. 'out#output_port' => 'in#input_port'),
all messages sent to all indexes of out#output_port will arrive at
in#input_port.

Note, this means that it is not possible to get messages sent to
"only" an unindexed output port.  Any connection will receive those
messages as well as all of the others.