Fixed a bug with keyed and unkeyed port interactions
When a component sends a message to a keyed port, it should also send
to any connected, unkeyed port.  For example, port['key'].send_message
should send the message to all connections for port['key'] and for
port[nil], which is what gets configured when you connect the port
without using a key.  Previously added all nil-keyed connections to
all keyed connection lists, however this is incorrect, as it only
works when the keyed ports are configured.  Chose to instead change
the [] method on the HashPort, which will add the nil-keyed ports to
the list that is returned, which has the right effect.  Had to add a
new method (add_connection), as we were previosly also using this []
method to then call << to add a connection.