Bind one-to-many connections on the cardinality-1 side.
Fan-in (many-to-one) worked already, because the consumer (input port)
`bind`ed the socket and all the producers (output ports) `connect`ed.

Fan-out (one-to-many) would only route to one consumer, because all the
consumers tried to `bind` the same socket, and all of them got clobbered
and the last one to `bind` was the only one to receive messages. We need
to run the `bind`/`connect` pair the other way for fan-out.

Many-to-many will require a broker solution using zmq_proxy.