Many-to-many connections spawn a broker process.
IPC sockets can only be bound once, but connected many times. To get
many pushers to spread their messages to many listeners, we start a
separate broker process that uses ZeroMQ devices to make the connection
(it binds on both sides and all producers and consumers connect).

Modern ZeroMQ code uses 'zeromq_proxy' as the call for this, but at this
time ffi-rzmq still maps the old Device API.