Maintain logger object identity throughout.
Previously, RFlow.logger was set to a startup logger while all
dependencies were loaded, then it was switched to the real logger. This
caused anyone who aliased RFlow.logger at load time to send their logs
permanently to the old startup log, which might be sent to /dev/null if
stdout was not attached and no log file was set. For example,
ActiveRecord::Base.logger was in this boat.

Now, RFlow.logger always maintains its identity, and when we switch over
the logger settings, we do it by rebuilding a new internal logger.

This has the happy side benefit of simplifying logger management in
bin/rflow and spec/spec_helper.

Fixes #56.