Move PID file management to DaemonProcess. (#78)
Move PID file management up so any process could have a PID file if we
wanted it. This may help us report PID file errors better as well,
because if we need to later we can handle the PID file writing from the
parent process. Right now, however, we don't need to, though we
originally thought we did. All the log messages all get into the log
just fine, so the real problem in #78 is that we need to also throw
something on the console from the parent process in case we're running
in background mode. That can be done without making this change, but
we're going to make this change anyway.